Title: 합과 평균의 차, Time: 45ms, Memory: 16MB, Status: Passed - CodeTree

+++ 250114/합과 평균의 차/sub-of-average-and-sum.py
@@ -5,4 +5,4 @@ a = int(arr[0])
 b = int(arr[1])
 c = int(arr[2])
 
-print((a+b+c), ((a+b+c)/3), (a+b+c)-((a+b+c)/3), sep = '\n')
+print((a+b+c), int((a+b+c)/3), int((a+b+c)-((a+b+c)/3)), sep = '\n')
